Move away from custom and incorrect integer types, just use proper ones
from <cstdint>. This avoids all the usual pain i'm having every time
i want to look at some snippet in godbolt.

Refs. https://github.com/darktable-org/rawspeed/issues/76

* int-type-cleanup:
  clang-tidy: finally enable google-runtime-int check
  Replace 'uint32' typedef with proper 'uint32_t' from <cstdint>
  Replace 'ushort16' typedef with proper 'uint16_t' from <cstdint>
  Replace 'uchar8' typedef with proper 'uint8_t' from <cstdint>
  Replace 'int32' typedef with proper 'int32_t' from <cstdint>
  Replace 'uint64' typedef with proper 'uint64_t' from <cstdint>
  Replace 'short16' typedef with proper 'int16_t' from <cstdint>
  Replace 'int64' typedef with proper 'int64_t' from <cstdint>
  Replace 'char8' typedef with proper 'int8_t' from <cstdint>

+++ src/librawspeed/common/Point.h
@@ -30,7 +30,7 @@ namespace rawspeed {
 class iPoint2D {
 public:
   using value_type = int;
-  using area_type = unsigned long long;
+  using area_type = size_t;
 
   constexpr iPoint2D() = default;
   constexpr iPoint2D(value_type a, value_type b) : x(a), y(b) {}
